+++ Sources/Views/ALKMyDocumentCell.swift
@@ -49,8 +49,6 @@ class ALKMyDocumentCell: ALKDocumentCell {
         super.setupViews()
 
         contentView.addViewsForAutolayout(views: [timeLabel, stateView])
-        stateView.widthAnchor.constraint(equalToConstant: Padding.StateView.width).isActive = true
-        stateView.heightAnchor.constraint(equalToConstant: Padding.StateView.height).isActive = true
         stateView.bottomAnchor.constraint(equalTo: bubbleView.bottomAnchor, constant: -Padding.StateView.bottom).isActive = true
         stateView.trailingAnchor.constraint(equalTo: bubbleView.leadingAnchor, constant: -Padding.StateView.trailing).isActive = true
 
@@ -65,26 +63,14 @@ class ALKMyDocumentCell: ALKDocumentCell {
 
     override func update(viewModel: ALKMessageViewModel) {
         super.update(viewModel: viewModel)
-
-        if viewModel.isAllRead {
-            stateView.image = UIImage(named: "read_state_3", in: Bundle.applozic, compatibleWith: nil)
-            stateView.tintColor = UIColor(netHex: 0x0578FF)
-        } else if viewModel.isAllReceived {
-            stateView.image = UIImage(named: "read_state_2", in: Bundle.applozic, compatibleWith: nil)
-            stateView.tintColor = nil
-        } else if viewModel.isSent {
-            stateView.image = UIImage(named: "read_state_1", in: Bundle.applozic, compatibleWith: nil)
-            stateView.tintColor = nil
-        } else {
-            stateView.image = UIImage(named: "seen_state_0", in: Bundle.applozic, compatibleWith: nil)
-            stateView.tintColor = UIColor.red
-        }
+        setStatusStyle(statusView: stateView, ALKMessageStyle.messageStatus)
     }
 
     override func setupStyle() {
         super.setupStyle()
         timeLabel.setStyle(ALKMessageStyle.time)
         bubbleView.backgroundColor = ALKMessageStyle.sentBubble.color
+        setStatusStyle(statusView: stateView, ALKMessageStyle.messageStatus)
     }
 
     class func heightPadding() -> CGFloat {