+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The changelog for [ApplozicSwift](https://github.com/AppLozic/ApplozicSwift). Al
 
 ### Enhancements
 - Added support for syncing package details when a suspension screen is shown.
+- Added an option to change message status(read receipt) style. Now icon for message status can be changed from outside or it can be hidden.
 
 ## [5.0.0] - 2020-02-14